+++ CHANGELOG.md
@@ -5,6 +5,14 @@ This file contains the notable changes in the ISLa project since version 0.2a1
 
 ## [unreleased]
 
+### Changed
+
+- Resolves [GitHub Issue #36](https://github.com/rindPHI/isla/issues/36), which in
+  essence refers to solving multiple "conflicting" SMT formulas. These are formulas
+  whose solutions change subtrees of substitutions in other SMT formulas. Before,
+  we did not check that the chosen "priority" formulas are not themselves influenced
+  by other formulas. See test case `test_issue_36` in `test_solver.py`.
+
 ## [1.9.9] - 2022-12-22
 
 ### Changed